Blazor Example updated for both Mock and EF core implementation
