+++ Samples/BlazorExample/BlazorExample/Server/Startup.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.ResponseCompression;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json.Serialization;
@@ -23,7 +24,14 @@ namespace BlazorExample.Server
                   new[] { "application/octet-stream" });
       });
 
-      services.AddTransient(typeof(DataAccess.IPersonDal), typeof(DataAccess.PersonDal));
+      //services.AddTransient(typeof(DataAccess.IPersonDal), typeof(DataAccess.PersonDal));
+      //for MockDb
+      // services.AddTransient(typeof(DataAccess.IPersonDal), typeof(DataAccess.Mock.PersonDal));
+
+      //for EF Db
+      services.AddTransient(typeof(DataAccess.IPersonDal), typeof(DataAccess.EF.PersonEFDal));
+      services.AddDbContext<DataAccess.EF.PersonDbContext>(
+      options => options.UseSqlServer("Server=servername;Database=personDB;User ID=sa; Password=pass;Trusted_Connection=True;MultipleActiveResultSets=true"));
       services.AddCsla();
     }
